[link-metrics] fix integer conversions & use of `u8` for offset
This commit addresses integer conversion warnings in `LinkMetrics`.
In particular it updates `ReadTypeIdFlagsFromMessage()` to use
`uint16_t` as offset value (instead of `uint8_t`).

+++ src/core/thread/link_metrics.hpp
@@ -433,8 +433,8 @@ private:
     Neighbor *GetNeighborFromLinkLocalAddr(const Ip6::Address &aDestination);
 
     static Error ReadTypeIdFlagsFromMessage(const Message &aMessage,
-                                            uint8_t        aStartPos,
-                                            uint8_t        aEndPos,
+                                            uint16_t       aStartPos,
+                                            uint16_t       aEndPos,
                                             Metrics &      aMetrics);
     static Error AppendReportSubTlvToMessage(Message &aMessage, uint8_t &aLength, const MetricsValues &aValues);
     static Error AppendStatusSubTlvToMessage(Message &aMessage, uint8_t &aLength, Status aStatus);
